@@ -13,7 +13,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3.0",
+        "php": ">=7.4.0",
         "firebase/php-jwt": "*"
     },
     "autoload": {
